fix warnings in windows build

+++ Source/Core/Core/HW/EXI/BBA/TAPServerBBA.cpp
@@ -80,7 +80,7 @@ void CEXIETHERNET::TAPServerNetworkInterface::HandleReceivedFrame(std::string&& 
   else
   {
     memcpy(m_eth_ref->mRecvBuffer.get(), data.data(), data.size());
-    m_eth_ref->mRecvBufferLength = data.size();
+    m_eth_ref->mRecvBufferLength = static_cast<u32>(data.size());
     m_eth_ref->RecvHandlePacket();
   }
 }
